move TIMEKEEPER functions into its own util
