+++ commands/drop.py
@@ -1,6 +1,7 @@
 from .common import *
 from utils.check_channel_access import *
 from utils.media_utils import *
+from utils.timekeeper import *
 
 command_config = config["commands"]["drop"]
 emojis = config["emojis"]
@@ -88,40 +89,3 @@ async def slash_drop(ctx:SlashContext, **kwargs):
       await ctx.send(f"{emojis.get('ezri_frown_sad')} Drop not found! To get a list of drops run: /drops", hidden=True)
   else:
     await ctx.send(f"{emojis.get('ohno')} Someone in the channel has already dropped too recently. Please wait a minute before another drop!", hidden=True)
-
-# Timekeeper Functions
-# Prevent spamming a channel with too many drops in too short a period
-#
-# TIMEKEEPER is a dict of tuples for each channel which have the last timestamp,
-# and a boolean indicating whether we've already told the channel to wait.
-# If we've already sent a wait warning, we just ignore further requests until it has expired
-TIMEKEEPER = {}
-TIMEOUT = 15
-
-async def check_timekeeper(ctx:SlashContext):
-  current_channel = ctx.channel.id
-
-  # Check if there's been a drop within this channel in the last TIMEOUT seconds
-  last_record = TIMEKEEPER.get(current_channel)
-  if (last_record != None):
-    last_timestamp = last_record[0]
-    diff = ctx.created_at - last_timestamp
-    seconds = diff.total_seconds()
-    if (seconds > TIMEOUT):
-      return True
-    else:
-      # Check if we've notified the channel if there's a timeout active
-      have_notified = last_record[1]
-      if (have_notified == False):
-        last_record[1] = True
-      return False
-
-  # If a timekeeper entry for the channel hasn't been set yet, go ahead and allow
-  return True
-
-
-def set_timekeeper(ctx:SlashContext):
-  current_channel = ctx.channel.id
-  TIMEKEEPER[current_channel] = [ctx.created_at, False]
-
-
